auth_server/requirements.txt
fastapi
uvicorn
sqlmodel
mysql-connector-python
cryptography
pyjwt
passlib[bcrypt]